# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AM Fleet API",
    "version": "0.1.9"
  },
  "server": {
    "url": "http://localhost:3333/",
    "description": "Local server"
  },
  "paths": {
    "/pecas": {
      "get": {
        "tags": [
          "pecas"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pecas"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pecas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pecas"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Peças criada",
            "headers": {
              "Location": {
                "description": "Link para peças criado",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/pecas/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "ID de peças",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "put": {
        "tags": [
          "pecas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pecas"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Peça alterada"
          },
          "404": {
            "description": "Peça não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Peça não encontrado"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pecas": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "nome": {
            "type": "string",
            "example": "Pneu"
          },
          "quantidade": {
            "type": "integer",
            "example": 12
          }
        },
        "required": [
          "nome",
          "quantidade"
        ]
      }
    }
  }
}